Moved UTF-8 code to its own class.
Adding comments, some renaming for clarity.
